API Cleanup - NAVBAR
